chore(auth): renaming
