+++ infra/dev/compose.dev.yaml
@@ -25,8 +25,8 @@ services:
     deploy:
       resources:
         limits:
-          cpus: '2.0'
-          memory: 2048M
+          cpus: '3.0'
+          memory: 3072M
         reservations:
           cpus: '0.5'
           memory: 512M
