+++ src/pages/signup/model/utils/field-name-mapper.ts
@@ -1,9 +1,9 @@
 import type { FieldPath } from 'react-hook-form';
-import { SignUpBody, SignupFormSchema } from '../schemas/SignupSchema';
+import { SignupBody, SignupFormSchema } from '../schemas/signup-schema';
 import { z } from 'zod';
 
 export const fieldNameMapper = (
-  fieldName: FieldPath<z.infer<typeof SignUpBody>>
+  fieldName: FieldPath<z.infer<typeof SignupBody>>
 ): FieldPath<z.infer<typeof SignupFormSchema>> => {
   switch (fieldName) {
     case 'firstName':
